Store workspace on server, sharable
